+++ app/rulesets/osu/performance/pp250306/evaluators/aim.go
@@ -1,7 +1,7 @@
 package evaluators
 
 import (
-	"github.com/wieku/danser-go/app/rulesets/osu/performance/pp250303/preprocessing"
+	"github.com/wieku/danser-go/app/rulesets/osu/performance/pp250306/preprocessing"
 	"github.com/wieku/danser-go/app/rulesets/osu/performance/putils"
 	"math"
 )
