+++ app/rulesets/osu/performance/pp250306/evaluators/flashlight.go
@@ -2,7 +2,7 @@ package evaluators
 
 import (
 	"github.com/wieku/danser-go/app/beatmap/difficulty"
-	"github.com/wieku/danser-go/app/rulesets/osu/performance/pp250303/preprocessing"
+	"github.com/wieku/danser-go/app/rulesets/osu/performance/pp250306/preprocessing"
 	"math"
 )
